Rename pages folder to avoid Streamlit's auto pages
