+++ app_pages/fund_scorecard.py
